pyproject.toml
[build-system]
requires = ["uv_build>=0.11.7"]
build-backend = "uv_build"


[project]
name = "ans-archiver"
version = "0.2.2"
description = "Web-scraper for downloading assignment submissions from the ANS (ans.app) assesment platform."
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "aiohttp>=3.13.1",
    "asyncio>=4.0.0",
    "bs4>=0.0.2",
    "color-parser-py>=0.1.6",
    "colorama>=0.4.6",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "yarl>=1.22.0",
]
[project.scripts]
    ans-archiver = "python.ans_submissions_archiver:main"
    ans-archiver-print = "python.printing_html_files:main"

[tool.pyright]
include = ["python/**/*.py"]

[dependency-groups]
dev = [
    "debugpy>=1.8.17",
]

[tool.uv.build-backend]
module-name = "python"
module-root=""